Refactor (rolling) : step2 - 프레임의 투구 결과를 별도의 클래스로 분리
