+++ src/main/java/bowling/domain/frame/FinalFrame.java
@@ -1,10 +1,12 @@
 package bowling.domain.frame;
 
+import bowling.domain.rolling.FinalRolling;
+
 public class FinalFrame implements  Frame {
-    private final FinalRollingResult rollingResult;
+    private final FinalRolling rollingResult;
 
     public FinalFrame() {
-        this.rollingResult = FinalRollingResult.init();
+        this.rollingResult = FinalRolling.init();
     }
 
     public void rollingBall(int pinCount) {
