+++ src/main/java/bowling/domain/rolling/FinalRolling.java
@@ -1,21 +1,23 @@
-package bowling.domain.frame;
+package bowling.domain.rolling;
 
 import bowling.domain.exception.InvalidThrowBallException;
 
-public class RollingResult {
+public class FinalRolling {
     private static final int KNOCKED_DOWN_PIN_COUNT_DEFAULT = -1;
     private static final int KNOCKED_DOWN_PIN_COUNT_STRIKE = 10;
 
     private Integer firstRollingResult;
     private Integer secondRollingResult;
+    private Integer thirdRollingResult;
 
-    private RollingResult() {
+    private FinalRolling() {
         this.firstRollingResult = KNOCKED_DOWN_PIN_COUNT_DEFAULT;
         this.secondRollingResult = KNOCKED_DOWN_PIN_COUNT_DEFAULT;
+        this.thirdRollingResult = KNOCKED_DOWN_PIN_COUNT_DEFAULT;
     }
 
-    public static RollingResult init() {
-        return new RollingResult();
+    public static FinalRolling init() {
+        return new FinalRolling();
     }
 
     public void roll(int pinCount) {
@@ -23,12 +25,17 @@ public class RollingResult {
             throw new InvalidThrowBallException();
         }
 
-        if ((firstRollingResult == KNOCKED_DOWN_PIN_COUNT_DEFAULT)) {
+        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_DEFAULT) {
             this.firstRollingResult = pinCount;
             return;
         }
 
-        this.secondRollingResult = pinCount;
+        if (secondRollingResult == KNOCKED_DOWN_PIN_COUNT_DEFAULT) {
+            this.secondRollingResult = pinCount;
+            return;
+        }
+
+        this.thirdRollingResult = pinCount;
     }
 
 
@@ -37,16 +44,28 @@ public class RollingResult {
             return false;
         }
 
-        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
-            return false;
-        }
-
         return true;
     }
 
     private boolean isFrameEnd() {
+        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
+            return true;
+        }
+
         if (firstRollingResult != KNOCKED_DOWN_PIN_COUNT_DEFAULT &&
                 secondRollingResult != KNOCKED_DOWN_PIN_COUNT_DEFAULT) {
+            return !shouldBonus();
+        }
+
+        return false;
+    }
+
+    private boolean shouldBonus() {
+        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
+            return true;
+        }
+
+        if (firstRollingResult + secondRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
             return true;
         }
