+++ src/main/java/bowling/domain/rolling/NormalRollings.java
@@ -1,23 +1,21 @@
-package bowling.domain.frame;
+package bowling.domain.rolling;
 
 import bowling.domain.exception.InvalidThrowBallException;
 
-public class FinalRollingResult {
+public class NormalRollings {
     private static final int KNOCKED_DOWN_PIN_COUNT_DEFAULT = -1;
     private static final int KNOCKED_DOWN_PIN_COUNT_STRIKE = 10;
 
     private Integer firstRollingResult;
     private Integer secondRollingResult;
-    private Integer thirdRollingResult;
 
-    private FinalRollingResult() {
+    private NormalRollings() {
         this.firstRollingResult = KNOCKED_DOWN_PIN_COUNT_DEFAULT;
         this.secondRollingResult = KNOCKED_DOWN_PIN_COUNT_DEFAULT;
-        this.thirdRollingResult = KNOCKED_DOWN_PIN_COUNT_DEFAULT;
     }
 
-    public static FinalRollingResult init() {
-        return new FinalRollingResult();
+    public static NormalRollings init() {
+        return new NormalRollings();
     }
 
     public void roll(int pinCount) {
@@ -25,17 +23,12 @@ public class FinalRollingResult {
             throw new InvalidThrowBallException();
         }
 
-        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_DEFAULT) {
+        if ((firstRollingResult == KNOCKED_DOWN_PIN_COUNT_DEFAULT)) {
             this.firstRollingResult = pinCount;
             return;
         }
 
-        if (secondRollingResult == KNOCKED_DOWN_PIN_COUNT_DEFAULT) {
-            this.secondRollingResult = pinCount;
-            return;
-        }
-
-        this.thirdRollingResult = pinCount;
+        this.secondRollingResult = pinCount;
     }
 
 
@@ -44,28 +37,16 @@ public class FinalRollingResult {
             return false;
         }
 
+        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
+            return false;
+        }
+
         return true;
     }
 
     private boolean isFrameEnd() {
-        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
-            return true;
-        }
-
         if (firstRollingResult != KNOCKED_DOWN_PIN_COUNT_DEFAULT &&
                 secondRollingResult != KNOCKED_DOWN_PIN_COUNT_DEFAULT) {
-            return !shouldBonus();
-        }
-
-        return false;
-    }
-
-    private boolean shouldBonus() {
-        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
-            return true;
-        }
-
-        if (firstRollingResult + secondRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
             return true;
         }
